Improve lazyload with events and wrapped functions
I need to wrap functions in the mappings, as otherwise the require is
being called on setup, meaning the plugin needs to be called on setup as
well.

When wrapping it inside a function, it will only be required when the
mapping being called, result in the possibility to lazy load it.