Remove VeryLazy event for vim plugins
I would like to keep it, but it results in the startup screen to
disappear. This is generally not problematic, but it annoys me and the
benefit of having these on `VeryLazy` does not really increase anything
really.

I might do this later on again and see or maybe I am doing something
wrong which results in this behaviour.