Lazy load surround and comment vim plugin
This will make these not useable in the first buffer that is opened, but
generally this is not needed. I mainly did this, because others might
also need to be active that are never really used (cmp for tab command
completion for example just works after insert enter currently).

So mainly this change is done to have a unified configuration. Maybe in
the future when I split even more I might revert this change (or if it
is not worth it) but for now this should be fine.