Add script to quickly copy the current git repo into a new folder
This is mainly meant as an alternative to git worktree.
Sometimes I need to quickly reference stuff inside another branch.
For example a PR. Instead of having to stash things and possibly
forgetting my previous work position etc. creating a copy of the repo
and switching that temporary file might be better and easier.

Then I can just work in this temporary folder for the PR (node_modules,
env files etc are all still there - and can be changed as wanted).
And remove it later on or re-copy it, when I need a newer version later
on.