Add trouble and run clippy automatically
Trouble enables easier workspace wide diagnostics in case the lsp
supports it (sadly typescript does not - but rust does).

It will also work with all open buffers, so at least something for lsps
without this feature.

The rust change will enable to run clippy on save and show general
improvements of code. Sadly I was not able to add the pedantic flags to
it (as I can not add them to the config file yet) but this is better
than nothing.