Replace `which-key` with general key search mapping
Generally I dont use which key at all and the popup just annoys me
whenever it pops up (because I think about some textobjects).

The only usecase I can see is that I forget a mapping, but for that I
added a telescope mapping and extended the description of the mappings
to make them better searchable.