Remove eslint and jest implementation for quickfix parsing
Sadly eslint and jest commands are not really well implemented if they
should be consumed by other cli applications (like nvim).
This results in them breaking too often or not working as they normally
work.

Additionally running them vie `npx` might be a root cause for some of
the problems with config files not being used etc. Overall it seems like
more of a waste of time then a nice to have thing (more headache then
niceness).

Additionally jest is really slow (thanks to the big repos I am using it
for) so using it there is not really that useful unless I add parameter
for specific files to run (which is overkill for me currently).

So instead of trying to fix them, I remove them again and maybe have a
look later on.