Refactor async vim command execution to handle outputs
I am used that I can use the shell to redirect stderr to stdin, as this
is more low level here and we are directly connected to both, I need to
allow the user to also say what to do on stderr.

So instead of having a general callback function, allow the user to
attach themselves to stdout or stderr in general (also ignoring if it
was not done via callback).

Generally using a class for some better passed table as well (ignoring a
lot of other options that could (and maybe should) be passed currently.

The output is currently always merged into one table, as I dont need to
use it any other way (might change later).