Replace oil.nvim with mini.files
I like oil.nvim but it seems like mini.files has most of the features
that I want with a bit of a better way of showcasing the filesystem.
I can still use the filesystem as a normal buffer, resulting in easy
move/delete/updates of the files that i want.