Remove language env in vim config
This is a basic bash thing, which should be done outside of vim.
Mainly a problem when the default language is non english, so the format is incorrect.
Sadly if i set it inside vim and this language is not supported, then it results in incorrect data as well.
So to prevent both, ill just leave this problem outside of vim for now.