Remove stopping lsp on big files
Main reason is, that stopping will stop the whole lsp making
the editor unusable if I want to continue with lsp in another buffer.

Big files should handle the lsp hopefully fine, but will see and at worst revert it.

Additionally I tried to detach the lsp, but I was not able to,as it is not attached
in this `lspattach` hook, and even if it seems to be reattaching otherwise