Disallow vim with all plugins to open big files
This also removes the generic check for treesitter, meaning that treesitter
is always on and not disabled.

Big files are normally files like history or some gigantic data files (excel, json)
these do not need many features of the IDE so instead of possibly crashing my current
vim setup I just not load them and notify the user to handle them differently.

It moves all these hanlding to the `vi` or minimal mode that has unperformant things disabled
resulting in no need to change the setup depending on given file (it will be depending
on usecase instead - decided by the user using `vim` or `vi`)