Update lazygit config
Fixe a performance problem (delta no dark default) as otherwise lazygit
will try to figure out which theme needs to be used on every call to
delta, which would be everytime a new commit or file is used.
This results in a big (2s) delay until getting the next diff to be
viewable.

Otherwise update some styling options and deprecated options