Remove commenting plugin
There is a basic version of commenting in the newest neovim version, as
I dont really use the current comment for special cases I can just
remove this.
This might be updated if I have files (vue) that still need a bigger
feature range, but for now just commenting lines and ranges are fine.