Move zshrc and ideavimrc to better fitting locations
There is no reason for zshrc to be the only config that is not in the
config folder.
It was there as the repo mimiced the linux dirs, but now its a bit more
split as I also dont use `.config` folders anymore.
The same for the ideavimrc, I move it to the `correct` path where it
belongs and only the linking script will know the real location at the
end.