Configure smart open to only use the cwd
Mainly as in the beginning it might show similar/same named files of
different repos.
For example searching for `Makefile` in my dotfiles also shows the
`Makefile` of my notes/other repos. This should be fixed over time as
the search gets smarter, but for now it is a possible result for
problems, so will filter them out.