Add script to forcefully reinstall neovim
Mainly as I use it often to install the latest neovim nightly instance,
instead of manually moving the old binary and calling the whole setup
script, which takes a few more seconds.
Now I can do all in one command which additionally is way faster to run,
as not all other parts are being executed as well.