Move custom command implementation and add keymap
I still want to keep it in a central place, so I can add more custom
commands from different places of my config (which is currently not
needed anymore as the TS command is gone).
Additionally add a keymap as the basic one with writing `CustomCommand`
is getting tedious if I want to keep using it.