Remap the quick selection keymap
Tab is my prefered key, but sadly it is internally mapped to CTRL-I
which is curcial for jumping.
It looks like this should work, but there might be some bugs in the
newest nightly, so Ill just leave it at that for now.

One potential fix would be to remap the CTRL-I to itself after I mapped
Tab, but that is more ugly then using my 2. mapping of just using leader
twice.