Ignore rename/action custom keybindings in vim
Mainly as these have now defaults in vim and I want to use these
instead. Before it just reminded me, but this was not obivous as the
normal action was still done.
Now it will just warn me and not do the action, so I need to do the new
keybind.