Try out a markdown highlighter
Mainly as I use markdown for notes, I want to see if this does anything
for me or if its just annoying and unhelpful.
It is still quite early in development, so I might want to keep it
around a bit to see if the vision is what I would like or not.