Add fish abbr to quickly edit root files with neovim
I mainly do not use sudoedit, as I could update it via
`SUDO_EDITOR=/bla/nvim` but it seems like I need to give the whole path,
which I dont want to for now.
It has some security problems (as now neovim will be run as sudo), but
ill leave it at that, as I rarely use it either way, and just need a
quick fix sometimes.