Create global gitignore file
Supermaven creates a config folder per project, so this is the first
time I really needed a global ignore as I dont want to always need to
update the gitignore of that profile.
This might be also useful for other things, but it should be kept in
mind, that only things that are "ONLY" for me should go in there, things
that others might also use (lets say vs code) should go into the
specified project, otherwise others will commit things without noticing
it (mainly as they just do `commit all` and never think about git).