Align all lsp specific keybindings to neovim defaults
grn, gra and grr are now the defaults and I want to use them instead of
my custom keybindings.
This should also be reflected in my other editors (but this is still
open for changes).

Additionally I removed `K` etc. from the config as these are set by
default. Signature is also already set by `CTRL-S` so ill remove it (as
I rarely use it either way).