Align formatting to the standard format expression of vim
Instead of using a custom mapping, we use the `gq` command. It seems
like the partial formatting does not work always, so I will mainly use
the `gQ` which is just the formatting of the whole file.

This should be fine as `gQ` opens ex-mode which I probably never need.