Add fallback for markdown keymaps
Sometimes I still want to use the original functionality of the key in
the markdown file, so I will fallback to it, when I can not do the
normal action either way.
This could also be done, by using a different key, but this will most
likely not work, as there are not many keys open, which ore single
digit.