Overwrite gq with formatting
Mainly as the command does not work most of the time with my custom
formatter, and even if, the setup of the `formatexpr` gets overwritten
most of the times.

Its rare that I want to format only a specific part, so thats fine for
me to format the whole file per default.