Ignore `-` in fish to improve script editing
I can edit command line commands via `ALT-e`, this will open the current
command in a vim buffer. This buffer is of type fish and has the `-` as
a word.
This might be fine, but is annoying as some bash commands could use to
be divided by `-`.
If this makes it worse in the long run, I can remove it again.