Align keymaps for minifiles better to normal vim usage
Allow me to use h and l to move instead of opening a file. This now can
be done via L and H instead.
Also I rarely used the <ENTR> to apply the things so using the normal =
is fine enought. Mapping the quitting to <ESC> is so that i can use the
q for macros.