Replace completion plugin cmp with blink
Mainly as cmp is not really worked on currently and because blink has
typo resistance, as this is one of my biggest problems with the current
completion.

Sadly cmdline and search does not have completion, so for these ill keep
cmp for now (might be that its irrelevant, then i can remove it
completly).

Also the speed should be better, but that I need to check.

One case that is obvious the completion for paths seems to be a bit
worse and the ordering is different (also filtering out things are not
always being done), but this is beta software so ill see how it goes