Align blink completion to my normal workflow
I normally write things, and do not want to be disturbed by the auto
completion all the time. So before when pressing enter it resulted in
the completion to be accepted.
Switching to use ESC to prevent this, is the same problem, as I need to
double ESC when i go from compltion to normal mode.

So the new way should be similar to my normal way, I show the completion
but without me selecting anything this does not do anything.
This also means that ESC will be used normally, and the completion is
just an opt-in and I need to manually say that I want to go into
"completion mode"