Swap back to cmp as a completion tool
Blink works fine, but it is missing auto import, it has no completion
for filter/comandline, breaks a bit too often, has some visual bugs and
in general the completion ordering is worse than cmp (stuff being
complete on the bottom, paths for example always being not usable etc).

Ill keep blink in with a config flag to test it in the future further.