Replace tmux with zellij as a terminal multiplexer
There are some things I can not do with zellij, but I want to try it out
and see if I find other things I like in zellij better.

One thing is, that I can not send commands in a hidden zellij session,
like in tmux. So making automation scripts might be more work.
As I currently dont use this as much (if at all), I dont mind ignoring
it for now and going without it.

In case I want to go back, look into this commit, as more than just one
place was updated (maybe an indicator that I need to change some setups
to not rely on specific tools).