Replace custom big file detection with snacks.bigfile approach
My previous implementation was a bit more explicit, I had a way to start
my neovim in a minimalistic mode, which makes opening big files
possible.

This approach is fine, but most of the time a bit annoying as this means
I need another session if there is a big file in a already open repo.
Additionally its not as good as it could be, as the implementation is
split into two, one to identify big files and make neovim not crash (NOT
that it can show it) and a way to start a new neovim instance that can
run it.

The new folke plugin snacks has another, more indirect approach. Instead
of handling this manually outside the editor, the files that are big get
flagged via the file type.
This allows us to use the filetype to disable all needed plugins
(treesitter/lsp are disabled by default).

So instead of having it manually be handled, now its done automatically
and even doable while the editor can still be used 100% in all other
contexts.

One note: the treesitter foldexpr was removed, as I did not find an easy
way to use it, and I dont use this in general, so dont see a need to
keep it.