Replace git helper functions in neovim with snacks
These functions do almost the same as before. The browser one is even
more integrated, so gitlab etc, should be even more usable.

The git blame one, should be almost 1:1, maybe a bit worse, but at least
it does not need external tmux, so im fine with it (normally i just need
to see the commit message, who did it and the hash to look further into
it, which are all provided