Improve debug-printing in vim via snacks
This allows all vim.print to use the inspect from snacks instead. this
is added to the notification and can be seen in the shownotification
command as well for quicker debugging.

There are other commands like backtrace etc, but i dont see the need
yet, so this is nice enought (even more as it will also work with `:=`
so instant lua execution)