Readd short overview on how to change foldtext
I removed the implementation as I dont use it, and it makes problems in
bigger files. As it took some getting used to, I wanted to readd it to
at least the utils file, in case I want to use it in the future again.

I am expecting this to completly get removed in the future, but thats
fine too.