Remove neovim blink completion plugin
As I do not use it, and it has missing features (command/search
completion), I will most likely not switch anytime soon.
Additionally there autocompletion order was weird when it comes to
pathing and other things, that just work with my cmp setup.

I might still switch in the future, but then I want to switch completly
and not have a half/half implementation. For that reason I remove it
here, to not keep old deprecated configurations - I will need to relearn
the plugin when its in a more mature state.