Replace dressing with snacks and fzf-lua
Mainly as dressing is not really maintained, and as snacks only has
something for input, fzf-lua seems to be a good selector.
The other way would be telescope, but ill try out fzf-lua maybe i could
switch back to it completly as a selector as I never had problems with
it.