Add vim alias to start without any options
Mostly as there still might be cases where the file has some weird stuff
which just breaks my config and makes it unusable.

For example big file does not handle if the file is quite big but its
all in one line. There highlight and others will quickly break.

Another option would be to let bigfile handle this, but as far as I can
see it only allows for the filesize, I would need something like "if one
line is soo long -> its a big file"