Tryout new lsp setup syntax from neovim 0.11
Sadly not all setups are easily done, so I still have some setups with
the lspconfig.
I want to find a good way to maintain this in general.

Also the new files are not able to run plugins, so these still need to
be inside the lsp.lua file.

I will have to do some trial and errors.