Remove ollama default installation
Depending on the OS I might want to install ollama in a different way.
Using wsl has limited features compared to just windows, when it comes
to gpu as a resource most importantly.

Additionally as AI models are resource heavy I probably dont want to
have ollama installed on every machine and will handle it case by case
depending on the machine itself.