Migrate to Mason v2 while removing automatic lsp/formatter/linter installation
I used the mason-lspconfig repo to map the different servers to the
installation path.
As now the mason-lspconfig has even less features (thanks to the new lsp
setup way), I dont want to continue using it.

This means the installation of lsps/formatters/linters etc. is being
done manually, which should be fine as this is a one-time thing either
way.

In case mason will provide an easy way to auto install things, I might
move back to it.