Show ignored files in grep/files picker
Because I sometimes want to edit .env files etc, which are ignored.
Additionally, its easier to toggle them off while they are on by
default.
An example could be, that I am not sure about the name, so I am
searching and never finding it, just then realizing that the ignored
files are hidden. The other way around is easier, as I will see too many
files that should have been ignored and then Ill just toggle in the ui.