Hide ignored files when searching files
Most of the time I am searching for some basic files which are also
versioned. I initially shown all, but when using it inside a repo with
node_modules or other generated files, the search is unusable.

So in the case I need to search for ignored files, I hope I can remember
to mnually switch it, or that the full text search is enough.