Remove winbar option in neovim
It sets the name of the window on the top left corner, I never really
used it as it would be nice when having splits, which I rarely have.
Otherwise its just the same data I have in my statusline.
Additionally it is confusing while using a picker, as the picker shows
the opened file over the previewed file, so no real need to keep it.