Handle command completions in vim for wsl
Seems like doing something like `:w !` will block the ui as the
completion is not being resolved at all. Mainly in wsl.

To prevent this but still have the rest of the completion, I just ignore
all cases that would use it.
It is only really needed in wsl, but as I hope it gets fixed in the
future, ill leave it in for all setups for now.