Prune git references on remote while fetching
This makes cleaning up after a successful merge request easier, as now
my local setup also has the info that the reference is not there anymore
and can therefore be quickly deleted.

The main benefit (next to cleaning up) is that creating new branches
with the same name will not result in potential conflicts between past
branches that are merged on remote but still have the reference to
origin locally.