Add mcp servers for accessing the browser and docs
One mcp allows the ai to open a browser and use it similar to a user.
This allows it to use websites that normally are not usable as they
require javascript to run.

The other helps with information about programming languages and should
extend the models trained information.

Additionally added an agent that forces itself to ask permissions for
everything that it can, still waiting to enforce asking for toolcalls as
well.