Disable opencode notification
The usage is weird. First I need the tab to be open, and then it
notifies my too often.

So either i get notified without needing it (as im actively looking or
dont need to know its done) or i miss that it completed and dont
continue with it as well.

Ill disable it for now but leave it in, in case I want to iterate on it.