Manually require all parts of my configuration
This makes the order explicit and allows me to more easily split the
configuration into sub folders etc.

As an example:
Completion needs to exist before lsp can start, as some configurations
depend on it. Either all needs to install and exist before setting it up
(similar to lazy.nvim) or I enforce the order and then manually handle
it.

This makes it a bit more tedious, but I want to see if there are other
drawbacks or if it just works fine