Remove unimportant mcp servers
It seems like with more mcps, the responses I get are getting worse. For
example when having rider active, it will als use riders tool to search
and replace text, even though there are perfectly working ones via the
terminal.

For now I want to limit it to these that are really worth it, until I
have another solution (maybe sub agents) that handle this better with
the same or better response quality.