Tryout catpuccin and move away from special markdown highlights
catpuccin had some weird highlights in markdown files, but otherwise
looks interesting and has many integrations that make other plugins look
more aligned with each other.

Because of the markdown part, I removed the explicit highlighting
plugin, which might not really do much for me anymore. It also helps
that the new theme kinda highlights markdown in a way, that already
solves most of my issues with it (completion state of todos)