Remove inc-rename
The feature is cool, but sadly it forces me to use the commandline,
which does not have the normal modes. So i can not easily jump in the
words, or delete specific parts etc.

This is currently possible because the input is being handled via
snacks, which is just a normal vim buffer.

There is a config to hijack this plugin, buts it looks hacky (has
commandline and snacks open at the same time). Therefore Ill ignore it
for now and hope it gets implenented in neovim itself in the future.