Remove unused custom treesitter foldtext
The basic one that just shows the line of the hidden text seems to be
better for myself.
The custom one shows only how many lines are hidden, this is also nice,
but I rather know what the initial line was and do not want to invest
more time into it for now.