Replace devicons via mini.icons
As I dont want mini part to be called in a specific order, I just rerun
the setup, which hopefully should be fine. Then at least each part has
the icons when needed and if I remove mini, ill see that its still being
required for the icons in other parts of my config.