Fix handling of locking hyprland
Hypridle needs to know how to trigger hyprlock, otherwise it can not
trigger it when the relevant events are being fired.
The same goes for my keybind that trigger the lock-session event. If
hypridle does not configure the lock command to hyprlock, then the event
will trigger nothing in the end.