Handle git work ignore explicitly
Before I ignored the folder but not 2 cases, this resulted in weird git
add warnings and vim was not able to find the folder (as it was
ignored).

Now I explicitly state that there might be a work git config file, and
that can be ignored. Means I need to add more ignored files if needed,
but so far its limited.