Replace notifier plugins with new neovim extui feature
This makes the `:messages` a real buffer and generally fixes the "press
enter" part.
So using it as a real notification way is possible, also has the benefit
of just using it with `:message` having `:messages clean` and keeping
the last message visible for longer than the popups by plugin.