Use wayland toolip value in json return
Instead of misusing the alt variable to then update the tooltip, it
seems like that i can set different values in the return json.

For example also the tooltip itself. I could do the same with "text" but
think its nicer like this to keep the formatting out of it where
possible (did not do it as the tooltip needs multiple values).