Switch to new rewritten nvim-treesitter plugin
This will only handle the parsers and installation. Meaning auto install
and enabling needs to be done manually.

Current setup will be a bit hacky on initial start, as the installation
and the enabling happens at the same time, but otherwise should be fine
enough.
There are most likely new features or ways of handling it in the future.